LPS-73562 Move _refreshBundles and _reinstallBundle out into an utility class
